Scroll to top moved for use in Gallery page, small tutorial edit in sanding page
